+++ backend/src/main/java/com/ssafy/forest/repository/ArticleCommentRepository.java
@@ -12,6 +12,6 @@ public interface ArticleCommentRepository extends JpaRepository<ArticleComment, 
 
     Page<ArticleComment> findAllByArticleOrderByCreatedAt(Pageable pageable, Article article);
 
-    int countArticleCommentByArticle(Article article);
+    long countArticleCommentByArticle(Article article);
 
 }
